reading strings instead of chars
But, things are getting messy. We have to read in substrings a lot and
iterate through them to malloc the right amount and to know how far to
jump ahead because we're looking at local copies of pointers. This is
less than ideal.

We've seen this pattern before- when you have too much stuff going on
because of making local copies, you can simplify things by passing
around a reference to the thing you're interested in, instead. So, a
pointer to a pointer OMG.