naming the enum brings benefits.
the compiler can now tell me if I have forgotten to handle a case in the
switch statement. this isn't a big deal now, but could become the source
of subtle bugs later on when I'm trying to eval things.

notice the neccesary `break;` statements at the end of each case.
Without break (or, alternately, as in the lower example in the read
switch,  a return statement), a switch will simply continue executing
the code in the following cases. it's really shorthand for `goto`
and the exit needs to be handled separately.